sentry tracking update, added NG_APP_SENTRY_RELEASE to support sentry relase
