+++ src/app/app.config.browser.ts
@@ -2,9 +2,6 @@ import { mergeApplicationConfig, ApplicationConfig, ErrorHandler } from '@angula
 import { appConfig } from './app.config';
 import * as Sentry from "@sentry/angular";
 
-function customErrorHandler(error: any) {
-  return error;
-}
 
 // @ts-expect-error i want this to compile, but it doesn't
 const SENTRY_DSN = import.meta?.env?.NG_APP_SENTRY_DSN;
@@ -14,15 +11,20 @@ const SENTRY_TRACE_SAMPLE_RATE = import.meta?.env?.NG_APP_SENTRY_TRACE_SAMPLE_RA
 const SENTRY_PROFILE_SAMPLE_RATE = import.meta?.env?.NG_APP_SENTRY_PROFILE_SAMPLE_RATE || '0';
 // @ts-expect-error i want this to compile, but it doesn't
 const SENTRY_ENVIRONMENT = import.meta?.env?.NG_APP_SENTRY_ENVIRONMENT || 'development';
+// @ts-expect-error i want this to compile, but it doesn't
+const SENTRY_RELEASE = import.meta?.env?.NG_APP_SENTRY_RELEASE || 'development';
 
 const BACKEND_BASE_URI = import.meta.env['NG_APP_BACKEND_BASE_URI'];
 
+
 let sentry_client: any = null;
+let providers = [];
 
 if(SENTRY_DSN && SENTRY_DSN !== '') {
   sentry_client = Sentry.init({
     dsn: SENTRY_DSN,
     environment: SENTRY_ENVIRONMENT,
+    release: SENTRY_RELEASE,
 
     integrations: [Sentry.browserTracingIntegration()],
 
@@ -36,15 +38,8 @@ if(SENTRY_DSN && SENTRY_DSN !== '') {
   });
 }
 
-let providers = [];
-
 if(SENTRY_DSN && SENTRY_DSN !== '') {
-  providers.push({
-    provide: ErrorHandler,
-    useValue: Sentry.createErrorHandler({
-      extractor: customErrorHandler
-    }),
-  },
+  providers.push(
   { provide: 'sentryClient', useValue: sentry_client });
 }
