+++ src/main.server.ts
@@ -11,13 +11,16 @@ const SENTRY_TRACE_SAMPLE_RATE = import.meta?.env?.NG_APP_SENTRY_TRACE_SAMPLE_RA
 const SENTRY_PROFILE_SAMPLE_RATE = import.meta?.env?.NG_APP_SENTRY_PROFILE_SAMPLE_RATE || '0';
 // @ts-expect-error i want this to compile, but it doesn't
 const SENTRY_ENVIRONMENT = import.meta?.env?.NG_APP_SENTRY_ENVIRONMENT || 'development';
+// @ts-expect-error i want this to compile, but it doesn't
+const SENTRY_RELEASE = import.meta?.env?.NG_APP_SENTRY_RELEASE || '1.0.0';
 
 const BACKEND_BASE_URI = import.meta.env['NG_APP_BACKEND_BASE_URI'];
 
 if(SENTRY_DSN && SENTRY_DSN !== '') {
   Sentry.init({
     dsn: SENTRY_DSN,
-    environment: SENTRY_ENVIRONMENT,
+    environment: SENTRY_ENVIRONMENT + 'server',
+    release: SENTRY_RELEASE,
 
     integrations: [Sentry.browserTracingIntegration()],
